refactor(web): featuresモジュールをFSDのセグメント構成へ再編
各スライス（authenticate / diary）の散在ファイルを ui / model / lib の
セグメントに分割し、スライスルートに Public API（index.ts）を新設。
スライス外からの参照は Public API 経由に統一した。

- ui: コンポーネント・Story
- model: hook・フォーム検証・型
- lib: 純粋なヘルパー（エラーメッセージ変換・Turnstile・集計）
