+++ application/packages/web/src/client/features/authenticate/index.ts
@@ -1,0 +1,13 @@
+export { AuthenticateForm } from './ui/authenticate-form'
+export {
+  AUTH_ERROR_MESSAGES,
+  resolveLoginErrorMessage,
+  resolveSignupErrorMessage,
+} from './lib/error-message'
+export { resolveTurnstileSiteKey } from './lib/turnstile'
+export { default as useSession, SESSION_SWR_KEY } from './model/use-session'
+export {
+  type AuthenticateErrors,
+  type AuthenticateFormData,
+  validateAuthenticateForm,
+} from './model/validation'
