+++ application/packages/web/src/client/routes/diary/analytics-page.tsx
@@ -8,17 +8,17 @@ import {
   ChartTooltip,
   ChartTooltipContent,
 } from '@/client/components/shadcn/chart'
-import DiaryLoginRequired from '@/client/features/diary/components/login-required'
-import DiaryPageLayout from '@/client/features/diary/components/page-layout'
-import DiaryReadListSection from '@/client/features/diary/components/read-list-section'
-import DiaryReadPagination from '@/client/features/diary/components/read-pagination'
-import DiarySummarySection from '@/client/features/diary/components/summary-section'
 import {
+  DiaryLoginRequired,
+  DiaryPageLayout,
+  DiaryReadListSection,
+  DiaryReadPagination,
+  DiarySummarySection,
   type ReadItem,
   type ReadPagination,
   type Source,
   type Summary,
-} from '@/client/features/diary/types'
+} from '@/client/features/diary'
 
 interface SummaryRangePoint {
   date: string
