+++ application/packages/web/src/client/routes/signup/hooks/use-signup.ts
@@ -3,12 +3,10 @@ import { useState } from 'react'
 import { useNavigate } from 'react-router'
 import {
   AUTH_ERROR_MESSAGES,
-  resolveSignupErrorMessage,
-} from '@/client/features/authenticate/error-message'
-import {
   type AuthenticateErrors,
+  resolveSignupErrorMessage,
   validateAuthenticateForm,
-} from '@/client/features/authenticate/validation'
+} from '@/client/features/authenticate'
 import getApiClientForClient from '@/client/infrastructure/api'
 
 export default function useSignup(turnstileSiteKey?: string) {
