+++ application/packages/web/src/client/routes/signup/route.tsx
@@ -1,5 +1,5 @@
 import { type LoaderFunctionArgs, type MetaFunction, useLoaderData } from 'react-router'
-import { resolveTurnstileSiteKey } from '@/client/features/authenticate/turnstile'
+import { resolveTurnstileSiteKey } from '@/client/features/authenticate'
 import useSignup from './hooks/use-signup'
 import SignupPage from './page'
